backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task queue
celery==5.4.0
redis==5.0.4

# Storage
boto3==1.34.84

# Config & validation
pydantic==2.7.1
pydantic-settings==2.2.1

# HTTP client
httpx==0.27.0

# Logging
structlog==24.1.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==10.3.0

# Testing
pytest==8.2.2
structlog==24.1.0
